+++ A2.txt
@@ -1,0 +1,4 @@
+Aufgabe2
+
+Bei 103.161.122.83 handelt es sich um eine IP-Adresse im Adressbereich des anggegebenen Netzes. Die 18. gibt an, dass es sich bei dein letzten 18. Bits um den Hostanteil und bei allen vorherigen Bits um den Netzanteil einer IP-Adresse im spezifizierten Netz handelt. Die Subnetzmaske berechnet man, indem man für jedes Bit des Netzanteils eine 1 und für jedes Bit des Hostanteils eine 0 setzt. In unserem Fall erhalten wir also die Subnetzmaske 255.252.0.0 (binär: 11111111.11111100.00000000.00000000). Die Netzwerkadresse erhält man aus der Verundung einer IP-Adresse im Netz und der Subnetzmaske (bzw. man setzt alle Bits im Hostanteil auf 0). In unserem Fall erhalten wir: 103.160.0.0 (binär: 01100111.10100000.00000000.00000000). Die Broadcastadresse erhält man aus der Veroderung einer IP-Adresse im Netz mit der negierten Subnetzmaske (bzw. man setzt alle Bits im Hostanteil auf 1): 103.163.255.255 (binär: 01100111.10100011.11111111.11111111).
+Der Adressbereich von 103.161.122.83/18 reicht von 103.160.0.0 (Netwerkadresse) bis 103.163.255.255 (Broadcastadresse). Somit überschneidet sich 103.161.193.83/18 mit diesem.
